fix typo + clarify needed change
Main reason: updated `CliError` made the code invalid.

Original code:
```
{{< high "rust" >}}
enum CliError {
    Io(io::Error),
    ParseInt(num::ParseIntError),
    ParseFloat(num::ParseFloatError),
}
{{< /high >}}

And add a new `From` impl:

{{< high "rust" >}}
impl From<num::ParseFloatError> for CliError {
    fn from(err: num::ParseFloatError) -> CliError {
        CliError::Parse(err)
    }
}
{{< /high >}}
```

```
        CliError::Parse(err)
```

Doesn't exist since the definition of `CliError` changed; this also invalidate the previous definition of `impl From<num::ParseIntError> for CliError`.  Therefore I changed the text to reflect that it's an update and provided the code change needed to reflect that.

That is
* update previous `impl From<num::ParseIntError> for CliError`
* add `impl From<num::ParseFloatError> for CliError`