Remove advice about failure
It's clear that failure isn't going anywhere. `anyhow`, with it's strict focus on dynamic errors and zero dependencies, actually seems like a good crate to recommend. Additionally, `std::error::Error` supports downcasting now just fine. It will support backtraces soon as well, but `anyhow::Error` still has an edge here, because it actually does capture the backtrace. 

However, it's probably premature to say that the advice in the blog is outdated, and will be replaced by X: so far, this blog post aged better than any of error-handing crates.