bstr: just use StdinLock
There's no need to wrap a StdinLock in a BufReader because StdinLock
already implements BufRead.

I had originally written a lot of this code using Stdin instead of
StdinLock, where Stdin does not implement BufRead.

But StdinLock is the right thing to use here.

Also, it is so much easier to use StdinLock now than it used to be. You
used to have to do something like this:

  let stdin = io::stdin();
  let stdin = stdin.lock();

Otherwise the lifetimes wouldn't work out. Turns out that we recently
made a change to Stdin::lock such that it returns a StdinLock<'static>
instead of a lifetime bound to Stdin. Which is so much nicer.

Ref: https://github.com/rust-lang/rust/pull/93965